requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
